Moving message-related helper methods to a MessageUtils class
